+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add @get_rss_from_block endpoint.
+  [cekk]
 
 
 0.1.0 (2020-04-08)